backend/requirements.txt
flask
Flask-SQLAlchemy
flask-login
flask-cors
flask-uploads
python-dotenv